adding component IntegrateLink
